feat: add routing for errors
