+++ packages/studio-web/src/app/app.component.html
@@ -1,79 +1,1 @@
-<mat-horizontal-stepper
-  linear
-  #stepper
-  (selectionChange)="selectionChange($event)"
->
-  <mat-step
-    [stepControl]="upload.uploadFormGroup"
-    i18n-label="Step 1"
-    label="Step 1"
-  >
-    <section>
-      <div class="container mb-2">
-        <div *ngIf="fatalError">
-          <h1 i18n="Language error heading" class="title">Aligner Error</h1>
-          <p i18n="Aligner error text">
-            Failed to initialize the aligner. This may be the result of an
-            incompatible browser. Please contact the developers and send them a
-            complete copy of the error message below:
-          </p>
-          <pre>{{ fatalError }}</pre>
-        </div>
-        <div *ngIf="!fatalError">
-          <h1 i18n="Welcome message for app" id="welcome-header">
-            Welcome to ReadAlong Studio
-          </h1>
-          <p i18n="Description of app">
-            This is a tool to help you make your own interactive 'readalong'
-            that highlights words as they are spoken. Have a look at
-            <a
-              href="https://www.eastcree.org/cree/en/lessons/read-along/northern-dialect/when-the-beaver-had-a-round-tail/"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              <mat-icon inline style="vertical-align: middle">launch</mat-icon>
-              this example in East Cree</a
-            >
-            to get a better idea of what these are.
-          </p>
-          <p i18n="Get started">
-            To get started making your own, read our
-            <a
-              style="color: #07c; cursor: pointer"
-              (click)="openPrivacyDialog()"
-              >privacy and data sovereignty policy</a
-            >, then take a tour by clicking on the tour button below, and follow
-            the steps below.
-          </p>
-          <div class="row">
-            <div class="col-3"></div>
-            <button
-              i18n="Take the tour"
-              class="mb-4 col-6 center text-center plausible-event-name=Tour"
-              mat-raised-button
-              color="primary"
-              (click)="startTour()"
-            >
-              Take the tour!
-            </button>
-            <div class="col-3"></div>
-          </div>
-        </div>
-      </div>
-    </section>
-    <app-upload #upload (stepChange)="stepChange($event)"></app-upload>
-  </mat-step>
-
-  <!-- <mat-step label="Step 2">
-        <app-config></app-config>
-    </mat-step> -->
-
-  <mat-step i18n-label="Step 2" label="Step 2">
-    <app-demo
-      *ngIf="b64Inputs$ | async as b64Inputs"
-      [b64Inputs]="b64Inputs"
-      [render$]="render$"
-    ></app-demo>
-  </mat-step>
-</mat-horizontal-stepper>
 <router-outlet></router-outlet>
