+++ packages/studio-web/src/app/fatal-error/fatal-error.component.html
@@ -1,0 +1,1 @@
+<p>fatal-error works!</p>
