+++ packages/studio-web/src/app/fatal-error/fatal-error.component.ts
@@ -1,0 +1,8 @@
+import { Component } from "@angular/core";
+
+@Component({
+  selector: "app-fatal-error",
+  templateUrl: "./fatal-error.component.html",
+  styleUrls: ["./fatal-error.component.sass"],
+})
+export class FatalErrorComponent {}
