+++ packages/studio-web/src/app/missing404/missing404.component.html
@@ -1,0 +1,1 @@
+<p>missing404 works!</p>
