+++ packages/studio-web/src/app/missing404/missing404.component.ts
@@ -1,0 +1,8 @@
+import { Component } from "@angular/core";
+
+@Component({
+  selector: "app-missing404",
+  templateUrl: "./missing404.component.html",
+  styleUrls: ["./missing404.component.sass"],
+})
+export class Missing404Component {}
