+++ packages/studio-web/src/app/studio/studio.component.sass
@@ -1,0 +1,4 @@
+#privacy
+    position: absolute
+    bottom: 10px
+    right: 10px
