+++ packages/studio-web/src/app/studio/studio.component.spec.ts
@@ -10,12 +10,12 @@ import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { RouterTestingModule } from "@angular/router/testing";
 
-import { AppComponent } from "./app.component";
-import { DemoComponent } from "./demo/demo.component";
-import { MaterialModule } from "./material.module";
-import { UploadComponent } from "./upload/upload.component";
+import { StudioComponent } from "./studio.component";
+import { DemoComponent } from "../demo/demo.component";
+import { MaterialModule } from "../material.module";
+import { UploadComponent } from "../upload/upload.component";
 
-describe("AppComponent", () => {
+describe("StudioComponent", () => {
   let httpClient: HttpClient;
   let httpTestingController: HttpTestingController;
   beforeEach(async () => {
@@ -29,15 +29,30 @@ describe("AppComponent", () => {
         ToastrModule.forRoot(),
         HttpClientTestingModule,
       ],
-      declarations: [AppComponent, UploadComponent, DemoComponent],
+      declarations: [StudioComponent, UploadComponent, DemoComponent],
     }).compileComponents();
     httpClient = TestBed.inject(HttpClient);
     httpTestingController = TestBed.inject(HttpTestingController);
   });
 
   it("should create the app", () => {
-    const fixture = TestBed.createComponent(AppComponent);
+    const fixture = TestBed.createComponent(StudioComponent);
     const app = fixture.componentInstance;
     expect(app).toBeTruthy();
+  });
+
+  it(`should have as title 'readalong-studio'`, () => {
+    const fixture = TestBed.createComponent(StudioComponent);
+    const app = fixture.componentInstance;
+    expect(app.title).toEqual("readalong-studio");
+  });
+
+  it("should have an h1", () => {
+    const fixture = TestBed.createComponent(StudioComponent);
+    fixture.detectChanges();
+    const compiled = fixture.nativeElement as HTMLElement;
+    expect(compiled.querySelector("h1")?.textContent).toContain(
+      "Welcome to ReadAlong Studio"
+    );
   });
 });
